pyproject.toml
[tool.poetry]
name = "bot0"
version = "0.1.0"
description = "Our zero-index Discord bot"
authors = ["Erik K Anderson <erik@ekand.io>"]
license = "CC0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.1"
python-dotenv = "^1.0.0"
pygithub = "^1.59.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
